updatesss

+++ audio/apis/global_db.py
@@ -4,6 +4,7 @@ import asyncio
 import contextlib
 import logging
 import urllib.parse
+from copy import copy
 
 from typing import TYPE_CHECKING, Mapping, Optional, Union
 
@@ -52,13 +53,15 @@ class GlobalCacheWrapper:
         if not self._token:
             self._token = await self.bot.get_shared_api_tokens("audiodb")
         self.api_key = self._token.get("api_key", None)
-        self.has_api_key = self.api_key is not None
+        self.has_api_key = self.cog.global_api_user.get("can_post")
         id_list = list(self.bot.owner_ids)
         self._handshake_token = "||".join(map(str, id_list))
         return self.api_key
 
     async def get_call(self, query: Optional[Query] = None) -> dict:
         api_url = f"{_API_URL}api/v2/queries"
+        if not self.cog.self.global_api_user.get("can_read"):
+            return {}
         try:
             query = Query.process_input(query, self.cog.local_folder_current_path)
             if any([not query or not query.valid or query.is_spotify or query.is_local]):
@@ -87,6 +90,8 @@ class GlobalCacheWrapper:
         return {}
 
     async def get_spotify(self, title: str, author: Optional[str]) -> dict:
+        if not self.cog.self.global_api_user.get("can_read"):
+            return {}
         api_url = f"{_API_URL}api/v2/queries/spotify"
         try:
             search_response = "error"
@@ -114,6 +119,8 @@ class GlobalCacheWrapper:
 
     async def post_call(self, llresponse: LoadResult, query: Optional[Query]) -> None:
         try:
+            if not self.cog.self.global_api_user.get("can_post"):
+                return
             query = Query.process_input(query, self.cog.local_folder_current_path)
             if llresponse.has_error or llresponse.load_type.value in ["NO_MATCHES", "LOAD_FAILED"]:
                 await asyncio.sleep(0)
@@ -147,7 +154,9 @@ class GlobalCacheWrapper:
     async def update_global(self, llresponse: LoadResult, query: Optional[Query] = None):
         await self.post_call(llresponse=llresponse, query=query)
 
-    async def report_invalid(self, id: str):
+    async def report_invalid(self, id: str) -> None:
+        if not self.cog.self.global_api_user.get("can_delete"):
+            return
         api_url = f"{_API_URL}api/v2/queries/es/id"
         async with self.session.delete(
             api_url,
@@ -155,3 +164,20 @@ class GlobalCacheWrapper:
             params={"id": id},
         ) as r:
             await r.read()
+
+    async def get_perms(self):
+        global_api_user = copy(self.cog.global_api_user)
+        await self._get_api_key()
+        if self.api_key is None:
+            return global_api_user
+        with contextlib.suppress(aiohttp.ContentTypeError, asyncio.TimeoutError):
+            async with aiohttp.ClientSession() as session:
+                async with session.get(
+                    f"{_API_URL}api/v2/users/me",
+                    headers={"Authorization": self.api_key, "X-Token": self._handshake_token},
+                ) as resp:
+                    if resp.status == 200:
+                        search_response = await resp.json()
+                        global_api_user.update(search_response)
+                        global_api_user["fetched"] = True
+        return global_api_user
